requirements.txt
elasticsearch
editdistance
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-project==0.9.1
analytics-python==1.4.0
anyio==3.5.0
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
asgiref==3.5.0
asn1crypto==1.4.0
astor==0.8.1
astroid==2.5
astropy==4.2.1
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.6
babel==2.9.0
backcall==0.2.0
backoff==1.10.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bert-score==0.3.11
bitarray==2.1.0
bkcharts==0.2
black==19.10b0
bleach==3.3.0
bokeh==2.3.2
boto==2.49.0
bottleneck==1.3.2
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
cloudpickle==1.6.0
clyent==1.2.2
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.4.7
cycler==0.10.0
cython==0.29.23
cytoolz==0.11.0
dask==2021.4.0
datasets==1.2.1
decorator==5.0.6
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.4
distributed==2021.4.1
docker-pycreds==0.4.0
docutils==0.17.1
entrypoints==0.3
et-xmlfile==1.0.1
faiss-cpu==1.7.2
fastapi==0.75.0
fastcache==1.1.0
ffmpy==0.3.0
filelock==3.0.12
flake8==3.9.0
flask==1.1.2
fonttools==4.31.2
frozenlist==1.3.0
fsspec==0.9.0
future==0.18.2
gevent==21.1.2
gitdb==4.0.9
gitpython==3.1.27
glob2==0.7
gmpy2==2.0.8
gradio==2.8.12
greenlet==1.0.0
h11==0.13.0
h5py==2.10.0
heapdict==1.0.1
html5lib==1.1
idna==2.10
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==3.10.0
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==5.3.4
ipython-genutils==0.2.0
ipython==7.22.0
ipywidgets==7.6.3
isort==5.8.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
jeepney==0.6.0
jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.4.1
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
jupyterlab-widgets==1.0.0
jupyterlab==3.0.14
keyring==22.3.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
libarchive-c==2.9
linkify-it-py==1.0.3
llvmlite==0.36.0
locket==0.2.1
lxml==4.6.3
markdown-it-py==2.0.1
markupsafe==1.1.1
matplotlib==3.5.1
mccabe==0.6.1
mdit-py-plugins==0.3.0
mdurl==0.1.0
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.1
mkl-service==2.3.0
mock==4.0.3
monotonic==1.6
more-itertools==8.7.0
mpmath==1.2.1
msgpack==1.0.2
multidict==6.0.2
multipledispatch==0.6.0
multiprocess==0.70.12.2
mypy-extensions==0.4.3
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
nltk==3.6.1
nose==1.3.7
notebook==6.3.0
numba==0.53.1
numexpr==2.7.3
numpy==1.20.1
numpydoc==1.1.0
olefile==0.46
openpyxl==3.0.7
orjson==3.6.7
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
paramiko==2.10.3
parso==0.7.0
partd==1.2.0
path==15.1.2
pathlib2==2.3.5
pathspec==0.7.0
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==8.2.0
pip==21.0.1
pkginfo==1.7.0
pluggy==0.13.1
ply==3.11
prettytable==2.1.0
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.17
protobuf==3.19.4
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyarrow==7.0.0
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.14.1
pycurl==7.43.0.6
pydantic==1.9.0
pydocstyle==6.0.0
pydub==0.25.1
pyerfa==1.7.3
pyflakes==2.2.0
pygments==2.8.1
pylint==2.7.4
pyls-black==0.4.6
pyls-spyder==0.3.2
pynacl==1.5.0
pyodbc==4.0.0-unsupported
pyopenssl==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pysocks==1.7.1
pytest==6.2.3
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-multipart==0.0.5
pytz==2021.1
pywavelets==1.1.1
pywsd==1.2.4
pyxdg==0.27
pyyaml==5.4.1
pyzmq==20.0.0
qdarkstyle==2.8.1
qtawesome==1.0.2
qtconsole==5.0.3
qtpy==1.9.0
regex==2021.4.4
requests==2.25.1
rope==0.18.0
rtree==0.9.7
ruamel-yaml-conda==0.15.100
sacremoses==0.0.49
scikit-image==0.18.1
scikit-learn==0.24.0
scipy==1.5.4
seaborn==0.11.1
secretstorage==3.3.1
send2trash==1.5.0
sentencepiece==0.1.96
sentry-sdk==1.5.8
setproctitle==1.2.2
setuptools==49.3.0
shortuuid==1.0.8
simplegeneric==0.8.1
singledispatch==0.0.0
six==1.15.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcollections==2.1.0
sortedcontainers==2.3.0
soupsieve==2.2.1
sphinx==4.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.4
spyder-kernels==1.10.2
spyder==4.2.5
sqlalchemy==1.4.15
starlette==0.17.1
statsmodels==0.12.2
sympy==1.8
tables==3.6.1
tblib==1.7.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
textdistance==4.2.1
threadpoolctl==2.1.0
three-merge==0.1.1
tifffile==2020.10.1
tokenizers==0.9.4
toml==0.10.2
tomli==2.0.1
toolz==0.11.1
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tornado==6.1
tqdm==4.49.0
traitlets==5.0.5
transformers==4.2.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
uc-micro-py==1.0.1
ujson==4.0.2
unicodecsv==0.14.1
urllib3==1.26.4
uvicorn==0.17.6
wandb==0.12.11
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==1.0.1
wheel==0.36.2
whoosh==2.7.4
widgetsnbextension==3.5.1
wn==0.9.1
wrapt==1.12.1
wurlitzer==2.1.0
xlrd==2.0.1
xlsxwriter==1.3.8
xlwt==1.3.0
xxhash==3.0.0
yapf==0.31.0
yarl==1.7.2
yaspin==2.1.0
zict==2.0.0
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.3.0